backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
supabase==2.3.7
httpx>=0.24.0,<0.26
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
apscheduler==3.10.4
python-multipart==0.0.9
Pillow>=10.0.0
numpy>=1.24.0
